Re-initial commit to raf's processing.js guide
"Don't judge it yet it's pre pre pre Alpha 0.01. I plan to add lots more content, nice markdown and hopefully humor. It took so long because I was trying to add stories and humor to it but it was so hard to get it right that I probably wrote and rewrote it five times before deciding to just do that after."