Link to workshops/README.md in main README.md
When you're linked to workshops/ on a small screen, the directory
listing fills up most of the screen, making it difficult to see what our
workshops are to a beginner.

This changes the workshop links from workshops/ to workshops/README.md
to make our list of workshops more understandable on smaller screens.