Fix spelling, grammar, and formatting errors
This change makes a ton of changes to the spelling, grammar, and
formatting in this repo to allow this repo to pass Markcop's new spell
check (see hackclub/markcop#50).

@paked deserves a huge shout-out for helping out with this. When we
started making these changes, Markcop found hundreds of errors. @paked
fixed nearly all of the errors in case_studies/ (100+ errors!) and made
an enormous impact on this change.