Remove canary
Hope that we can re-add this once we have a little more bandwidth. We
haven't received a NSL or orders under FISA.

I'm not publishing a PGP signed message right now because I don't
currently have physical access to a computer that has my PGP key due to
current location.