Removing link to Molleindustria
Molleindustria is blocked for pornographic content in my district for some obvious reasons(look at their homepage). If it's blocked on ours for that reason, it might be blocked in other districts as well. It might be a good idea to remove the link to Molleindustria to remove the possibility of students getting in trouble for attempting to access this website, along with preventing the potential association of pornography to Hack Club.