Swap Finding Bigfoot and That Was Easy
Did a couple of things here! i added the `finding bigfoot` workshop to the "starter group" (and added my name to the authors because of the extension I wrote, if that's okay) and then took out the `The Was Easy` workshop and put it in experimental. I also re numbered the experimental groups because the meta data was out of order