Criteria:
 Workshop that meets all of the "Good workshops" guidelines outlined above (View below)
 1-2 minute video clip from your test club meeting showing what members built (View: https://hackclub.slack.com/files/U013CQTVBNZ/F01H6BGJ5FH/hackclub_flappy_bird_demo.mp4)
 All assets (images, PDFs, and any other binaries) must be hosted using the #cdn channel on the Hack Club Slack for posterity and to cut down on repository size
 All workshops must be in English at native-level proficiency. Phrasing and grammar must be perfect.
 You must be either a registered Hack Club leader or an active member of the Slack (I am a club leader, and I also count as an active member in the community, I hope)
Good workshops guidelines:

Help club members produce a compelling and real project in under 20 minutes

This workshop can be completed in under 20 minutes, based on the test meeting. I would define this as a compelling project as it is something they can show off and it is very visual. It is a real project as they will be using HTML, CSS & jQuery to make a beatmaker/drum pad.

Are very easy to hack on for long periods of time after completion

Yes, during the test meeting only people found it interesting and wanting to add more functionality like recording etc. And having future plans with the project.

Have instructive and compelling images, GIFs, and visuals

It does have them.

Use repl.it or a similar service that allows members to begin coding with one click

Yes.

Link unknown concepts to external sources.
Yes, I have linked docs and unknown sources.

It relates CSS animation keyframing to video animation keyframing and relates CSS variables to how variables in other languages.

End with at least 3 demos of how the project can be expanded, with live links and code links for each demo.

Yes, it has 3 expansions on the project with live links code links.

Please let me know if I missed anything, and feel free to improve it.