Javascript Typing Test (workshop)
Hello there! This is my second pull request, so obviously I have to thank you for the dedication you put into reviewing this for me.

The link to the mock meeting:
https://hackclub.slack.com/archives/C01A9V5LPD0/p1609467151063400

We create a small little typing speed test in this workshop. It tells you the WPM, accuracy, and time.