Switch time statement in code
In the demo video, `time.sleep(1)` comes before `tweetButton.click()`, but it was the other way around on the workshop page. I got an error message when I copy/pasted the code from the workshop page. Therefore, I concluded that the code in the demo video was correct and the code on the page was not.