Fixed inputRange formatting.
inputRange was written incorrectly as inputrange, with a lower-cased "r", which caused some problems when trying to replicate the workshop project. By changing the lower-cased "r" with an upper-cased "R", bugs/conflicts were resolved.

This edit was made based on @wollygfx's provided demo/source-code for this project (https://repl.it/@wollygfx/2-Image-editor).